Update substring parser to comply with Bash's rules
The substring parser is a Bash extension to POSIX Parameter Expansion, and it turns out our parsing has been different from how Bash's works for a while now, causing the singular gotcha that it is impossible to substitute a numeric value in the default value expansion.

For example, this expansion: `${EMPTY:-127}` works in Bash and Zsh, giving you a result of `127` if `EMPTY` is unset, however, the same expansion is treated as a negative-offset substring expansion by interpolate, meaning you end up with a blank value rather than the expected default.

This seems to have been written this way intentionally, but it seems to be an oversight.

I've updated the parser to not special-case a numeric character following the :- operator, as the Bash-approved way is to require a space character after the - to disambiguate the two similar, but distinct functions.