Include escaped expansions in calls to Identifiers
It's useful for consumers of this libary to be able to know if their input contains things that would be interpolated, and if so, which things. Currently, this is handled by the `interpolate.Identifiers()` function, which returns a list of the interpolate-able identifiers that `interpolate.Interpolate()` would replace.

However, this library makes one further modification to its input: when escaped expansions are included (ie, $$MY_VAR or \$MY_VAR), they're de-escaped in the output. This is working as intended, but the `interpolate.Identifiers` function doesn't include these escaped interpolations, which is a bit confusing when you're using the `Identifiers` function to determine "will running Interpolate() change the input?"

To remedy this, this PR updates the parse to handle escaped interpolations a bit differently, and treat them as a kind of expansion that doesn't take any input, and de-escapes its input. There's no functionality change, but escaped interpolations now show up in the `Identifiers()` function